+++ src/BuildStage.groovy
@@ -1,0 +1,10 @@
+public class BuildStage implements Stage {
+    @Override
+    public Closure pipelineConfiguration() {
+        return { ->
+            stage("build") {
+                sh("./bin/build.sh")
+            }
+        }
+    }
+}