+++ test/JenkinsfileTest.groovy
@@ -1,0 +1,11 @@
+import static org.hamcrest.MatcherAssert.assertThat
+import static org.hamcrest.Matchers.equalTo
+
+import org.junit.jupiter.api.Test
+
+class JenkinsfileTest {
+    @Test
+    void foo() {
+        assertThat(true, equalTo(true))
+    }
+}